+++ Releases.md
@@ -35,7 +35,7 @@ For the time being, these are some of the CTF Challenges intended to be relased 
 | Meow               | HackTheBox, Network Services, Completed                          | Done     |               |
 | Daily Bugle        | TryHackMe, SQLmap                                                | Done     |               |
 | Thompson           | Reverse Shell, Network Services, TryHackMe                       | Exported | 03/11/23      |
-| Library            | TryHackMe, Network Services                                      | Done     |               |
+| Library            | TryHackMe, Network Services                                      | Exported | 06/11/23      |
 | Oh My WebServer    | TryHackMe                                                        | Done     |               |
 | tomghost           | TryHackMe, Boot2Root                                             | Done     |               |
 | Anonforce          | TryHackMe, Boot2Root, Network Services                           | Done     |               |
@@ -70,3 +70,25 @@ For the time being, these are some of the CTF Challenges intended to be relased 
 | RootMe             | CTF, TryHackMe                                                   | Exported | 26/10/23      |
 | Pickle Rick        | TryHackMe, CTF                                                   | Done     |               |
 | RES                | TryHackMe, CTF                                                   | Done     |               |
+| Codify             | HackTheBox                                                       | Exported | 06/11/23      |
+
+
+
+## Coding Challenges
+Every Monday,Wednesday and Friday, new Coding Challenge Solutions are going to be posted, where ;
+1. JavaScript :: Monday
+2. Python :: Wednesday
+3. Friday :: PHP
+
+| Challenge Title | Description | Link |
+| --------------- | ----------- | ---- |
+| Encrypt This! | Given a string, this function encrypts it by replacing the first letter of each word with its character code and rearranging the rest of the letters. | [Challenge 1](https://www.codewars.com/kata/5848565e273af816fb000449/javascript) |
+| Make the Deadfish Swim | This function interprets a string of commands ('i', 'd', 's', 'o') to manipulate a value and returns an array of results based on the commands. | [Challenge 2](https://www.codewars.com/kata/51e0007c1f9378fa810002a9) |
+| Pyramid Array | This function generates a pyramid-shaped array of 1s. | [Challenge 3](https://www.codewars.com/kata/515f51d438015969f7000013) |
+| Create Phone Number | Given an array of 10 digits, this function formats it into a phone number. | [Challenge 4](https://www.codewars.com/kata/525f50e3b73515a6db000b83) |
+| Calculate Pyramid Height | This function calculates the height of a pyramid given the number of blocks used. | [Challenge 5](https://www.codewars.com/kata/56968ce7753513604b000055) |
+| Detect Pangram | This function checks if a given string is a pangram (contains all letters of the alphabet). | [Challenge 6](https://www.codewars.com/kata/545cedaa9943f7fe7b000048) |
+| Roman Numerals Decoder (Default Solution) | This function decodes a Roman numeral string and converts it into an integer. | [Challenge 7](https://www.codewars.com/kata/51b6249c4612257ac0000005) |
+| Roman Numerals Decoder (Alternative Solution) | This function provides an alternative solution to decode Roman numerals. | [Challenge 7](https://www.codewars.com/kata/51b6249c4612257ac0000005) |
+| Tim's Fruit Shop Challenge | This function optimizes the packaging of fruit orders into bags, boxes, and pallets based on the number of units. | [Challenge 8](https://www.codewars.com/kata/652643925c042100247fffc6/) |
+